Develop

+++ CBP/library/api/fermat-cbp-api/src/main/java/com/bitdubai/fermat_cbp_api/layer/wallet/crypto_broker/interfaces/CryptoBrokerStockTransactionRecord.java
@@ -5,15 +5,17 @@ import com.bitdubai.fermat_cbp_api.all_definition.enums.CurrencyType;
 import com.bitdubai.fermat_cbp_api.all_definition.enums.TransactionType;
 import com.bitdubai.fermat_cbp_api.all_definition.wallet.StockTransaction;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
  * Created by Yordin Alayn on 02.10.15.
  */
 public interface CryptoBrokerStockTransactionRecord extends StockTransaction {
+    //TODO: Documentar
 
-    float getRunningBookBalance();
+    BigDecimal getRunningBookBalance();
 
-    float getRunningAvailableBalance();
+    BigDecimal getRunningAvailableBalance();
 
 }
